Updated support form z index, button location on mobile. Resolves #729.
