+++ src/components/layout.module.css
@@ -18,26 +18,3 @@
 	position: fixed;
 	right: 0;
 }
-
-/* Support request button */
-.supportRequestBtn {
-	background-color: var(--accent);
-	border: 0;
-	border-top-left-radius: 2px;
-	border-top-right-radius: 2px;
-	color: var(--white);
-	cursor: pointer;
-	outline: 0;
-	padding: 8px 16px;
-	position: fixed;
-	right: 0;
-	top: 40%;
-	transform:rotate(-90deg);
-	transform-origin: right bottom;
-	transition: background-color 500ms cubic-bezier(0.25, 0.8, 0.25, 1);
-	z-index: 100;
-}
-
-.supportRequestBtn:hover {
-	background-color: var(--accent-hover);
-}
